move some functions into helper file
